requirements.txt
anyio==4.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing-extensions==4.13.2
urllib3==2.2.3
